+++ openhands_cli/listeners/__init__.py
@@ -1,0 +1,5 @@
+from openhands_cli.listeners.pause_listener import PauseListener
+
+__all__ = [
+    "PauseListener",
+]
